Better support for VirtualBox
This fixes a problem with VirtualBox where eth0 is the NAT network, which cannot access other Consul servers. 
This *hack* uses eth1 first if it's available as the bind address. This then connects to the Consul servers using the private IP. 

I don't think this should go directly in to the code line because it changes existing behavior, but it's a useful change for those using VirtualBox.